requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
asyncpg==0.29.0
SQLAlchemy==2.0.31
boto3==1.34.162
requests==2.32.3
python-dotenv==1.0.1
torch==2.3.1
diffusers==0.30.0
transformers==4.42.3
accelerate==0.33.0
pillow==10.4.0
huggingface_hub==0.24.5
pytest==8.3.2
pytest-asyncio==0.23.8
